输出mango内部的状态

+++ src/main/java/cc/concurrent/mango/benchmark/model/Stat.java
@@ -28,7 +28,7 @@ public class Stat {
         return totalCost;
     }
 
-    public long getAvg() {
+    public double getAvg() {
         return totalCost / (successNum + exceptionNum);
     }
